Fixed bugs and optimized forwardIndexGenerator to take 81 hours instead of 10 weeks

+++ forwardInvertedIndex.sql
@@ -2,7 +2,7 @@ create table forwardIndex
 (
   docID int,
   wordID int,
-  nHits  tinyint unsigned,
+  nHits  smallint unsigned,
   hit smallint unsigned
 );
 
@@ -11,7 +11,7 @@ create table invertedIndex
   wordID int,
   nDocs int,
   docID int,
-  nHits tinyint unsigned,
+  nHits smallint unsigned,
   hit smallint unsigned
 );
